Export from Fitbit studio v0.0.1.
This allows the user to swipe up or down to select a dice, then swipe
left or right to make a roll. Swipe any direction to return to the dice
and select a new dice or make a new roll.

The dice set is from https://www.iconspng.com/image/8526/rpg-dice